requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0
